+++ src/types/mod.rs
@@ -15,11 +15,10 @@ pub use binary::MqttBinary;
 pub use int::VarByteInt;
 pub use qos::{IdentifiedQoS, QoS};
 pub use reason_code::ReasonCode;
-pub use string::MqttString;
+pub use string::{MqttString, MqttStringError};
 pub use topic::{TopicFilter, TopicName};
 
 /// Variable byte integer: If `VarByteInt::MAX_ENCODABLE` is exceeded, an error of this type is returned.
-/// MQTT string: If `MqttString::MAX_LENGTH` is exceeded, an error of this type is returned.
 /// MQTT binary: If `MqttBinary::MAX_LENGTH` is exceeded, an error of this type is returned.
 #[derive(Debug)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]