+++ src/v5/packet/pings/mod.rs
@@ -45,7 +45,7 @@ impl<'p, T: PingPacketType> RxPacket<'p> for GenericPingPacket<T> {
 impl<T: PingPacketType> TxPacket for GenericPingPacket<T> {
     fn remaining_len(&self) -> VarByteInt {
         // Invariant: 0 < VarByteInt::MAX_ENCODABLE
-        VarByteInt::new(0)
+        VarByteInt::new_unchecked(0)
     }
 
     async fn send<W: Write>(&self, write: &mut W) -> Result<(), TxError<W::Error>> {