+++ src/v5/packet/unsubscribe.rs
@@ -86,7 +86,7 @@ impl<'p, const MAX_TOPIC_FILTERS: usize> UnsubscribePacket<'p, MAX_TOPIC_FILTERS
 
     pub fn properties_length(&self) -> VarByteInt {
         // Invariant: Max length = 0 < VarByteInt::MAX_ENCODABLE
-        VarByteInt::new(0)
+        VarByteInt::new_unchecked(0)
     }
 }
 
@@ -106,10 +106,10 @@ mod unit {
         let mut topics = Vec::new();
 
         topics
-            .push(TopicFilter::new_checked(MqttString::try_from("test/+/topic").unwrap()).unwrap())
+            .push(TopicFilter::new(MqttString::try_from("test/+/topic").unwrap()).unwrap())
             .unwrap();
         topics
-            .push(TopicFilter::new_checked(MqttString::try_from("test/#").unwrap()).unwrap())
+            .push(TopicFilter::new(MqttString::try_from("test/#").unwrap()).unwrap())
             .unwrap();
 
         let packet: UnsubscribePacket<'_, 2> = UnsubscribePacket::new(9874, topics).unwrap();